requirements.txt
langchain
langchain-openai
langchain-community
chromadb
neo4j
pypdf
python-dotenv
requests
streamlit
